cleanup: cleans up API surface and fixes integration test deadlock
